+++ work/Installer/Build-Release.ps1
@@ -1,7 +1,7 @@
 [CmdletBinding()]
 param(
     [ValidatePattern('^\d+\.\d+\.\d+$')]
-    [string]$Version = '0.5.0',
+    [string]$Version = '0.6.0',
     [string]$DevenvPath,
     [string]$OutputDirectory,
     [switch]$PublishToGitHub,